port_code/requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pupil-apriltags==1.0.4.post10
RPi.GPIO==0.7.1
